+++ lib/multiple_result.dart
@@ -1,5 +1,6 @@
 library multiple_result;
 
 export 'src/async_result.dart';
+export 'src/function.dart';
 export 'src/result.dart';
 export 'src/unit.dart';